Add DoubleType ReverseArray Method
