+++ src/utils/ReverseArray.java
@@ -1,0 +1,27 @@
+package utils;
+
+public class ReverseArray {
+
+    public static int[] reverse(int[] array){
+
+        int[] reverse = new int[array.length];
+
+        for (int i = array.length - 1, j=0; i >= 0; i--) {
+            reverse[j++]= array[i];
+        }
+
+        return reverse;
+    }
+
+    public static double[] reverse(double[] array){
+
+        double[] reverse = new double[array.length];
+
+        for (int i = array.length - 1, j=0; i >= 0; i--) {
+            reverse[j++]= array[i];
+        }
+
+        return reverse;
+    }
+
+}
