--- src/utils/ReverseString.java
@@ -1,16 +1,0 @@
-package utils;
-
-public class ReverseString {
-
-    public static int[] reverse(int[] array){
-
-        int[] reverse = new int[array.length];
-
-        for (int i = array.length - 1, j=0; i >= 0; i--) {
-            reverse[j++]= array[i];
-        }
-
-        return reverse;
-    }
-
-}
